Basic job system in place. Right click to send movement jobs and see workers go to the job location

+++ lib/SXNGN/headers/ECS/Utilities/Map_Utils.hpp
@@ -22,7 +22,7 @@ namespace SXNGN {
 
 				static std::queue<Location> GetPath(NAVIGATION_TYPE method, Location start, Location end);
 				
-				static std::pair<int, int> GetVector(Location start, Location end);
+				static std::pair<double, double> GetVector(Location start, Location end);
 
 			};
 		}
